Move how to join and program benefits out
