+++ src/components/pages/educators/index.js
@@ -1,3 +1,5 @@
+import HowToJoin from './how-to-join';
+import ProgramBenefits from './program-benefits';
 import SignUpModal from './sign-up-modal';
 
-export { SignUpModal };
+export { HowToJoin, ProgramBenefits, SignUpModal };
